requirements.txt
black==22.8.0
Django~=3.2.7
django-rest-swagger==2.2.0
djangorestframework==3.12.2
djangorestframework-filters==0.10.2
flake8==4.0.1
gevent==21.1.2
gunicorn==20.0.4
mkdocs==1.1.2
mkdocs-material==7.0.3
pre-commit==2.12.1
pylint==2.8.2
pylint-django==2.4.4
pytest==7.1.3
pytest-xdist==2.5.0
pytest-cov==4.0.0
python-decouple==3.4
psycopg2==2.9.1
requests==2.20.0
